Add header templates
Added header home and header library templates, added styles and logic
for change header template
