+++ src/partials/header.html
@@ -1,42 +1,19 @@
 <header class="header">
   <div class="container">
-    <nav class="logo">
+    <a class="logo link" href="">
       <img class="logo-img" src="./images/logo.png" alt="logo" />
       <p class="logo-text">Filmoteka</p>
-    </nav>
-
+    </a>
     <nav class="header-menu">
       <ul class="header-list list">
         <li class="header-item">
-          <a class="header-link link active" href="">Home</a>
+          <a class="header-link link-home-js link active" href="#">Home</a>
         </li>
         <li class="header-item">
-          <a class="header-link link" href="">my library</a>
+          <a class="header-link link-library-js link" href="#">my library</a>
         </li>
       </ul>
     </nav>
-    <div class="header-input-wrapped">
-      <label class="header-input-label" for="">
-        <input
-          id="header-input"
-          class="header-input"
-          type="text"
-          placeholder="Поиск фильмов"
-        />
-        <svg class="header-input-icon">
-          <use href="./images/sprait/symbol-defs.svg#icon-search-2"></use>
-        </svg>
-      </label>
-    </div>
-    <!-- <div class="header-button-wrapped">
-      <ul class="header-button-list list">
-        <li class="header-button-item item">
-          <button class="button button-header">Watched</button>
-        </li>
-        <li class="header-button-item item">
-          <button class="button button-header">queue</button>
-        </li>
-      </ul>
-    </div> -->
+    <div class="header-content"></div>
   </div>
 </header>
