+++ src/templates/header-home.hbs
@@ -1,0 +1,13 @@
+<div class='header-input-wrapper'>
+  <label class='header-input-label' for=''>
+    <input
+      id='header-input'
+      class='header-input'
+      type='text'
+      placeholder='Поиск фильмов'
+    />
+    <svg class='header-input-icon'>
+      <use href='./images/sprait/symbol-defs.svg#icon-search-2'></use>
+    </svg>
+  </label>
+</div>
