+++ src/templates/header-library.hbs
@@ -1,0 +1,10 @@
+<div class='header-button-wrapper'>
+  <ul class='header-button-list list'>
+    <li class='header-button-item item'>
+      <button class='button button-header'>Watched</button>
+    </li>
+    <li class='header-button-item item'>
+      <button class='button button-header'>queue</button>
+    </li>
+  </ul>
+</div>
